+++ scenes/main_menu.c
@@ -5,6 +5,8 @@ enum {
     MainMenuAutoDetect,
     MainMenuHW3,
     MainMenuHW4,
+    MainMenuLegacy,
+    MainMenuSettings,
     MainMenuAbout,
 };
 
@@ -21,6 +23,8 @@ void tesla_fsd_scene_main_menu_on_enter(void* context) {
     submenu_add_item(app->submenu, "Auto Detect & Start", MainMenuAutoDetect, main_menu_callback, app);
     submenu_add_item(app->submenu, "Force HW3 Mode", MainMenuHW3, main_menu_callback, app);
     submenu_add_item(app->submenu, "Force HW4 Mode", MainMenuHW4, main_menu_callback, app);
+    submenu_add_item(app->submenu, "Force Legacy Mode", MainMenuLegacy, main_menu_callback, app);
+    submenu_add_item(app->submenu, "Settings", MainMenuSettings, main_menu_callback, app);
     submenu_add_item(app->submenu, "About", MainMenuAbout, main_menu_callback, app);
 
     view_dispatcher_switch_to_view(app->view_dispatcher, TeslaFSDViewSubmenu);
@@ -46,6 +50,16 @@ bool tesla_fsd_scene_main_menu_on_event(void* context, SceneManagerEvent event) 
             app->hw_version = TeslaHW_HW4;
             fsd_state_init(&app->fsd_state, TeslaHW_HW4);
             scene_manager_next_scene(app->scene_manager, tesla_fsd_scene_fsd_running);
+            consumed = true;
+            break;
+        case MainMenuLegacy:
+            app->hw_version = TeslaHW_Legacy;
+            fsd_state_init(&app->fsd_state, TeslaHW_Legacy);
+            scene_manager_next_scene(app->scene_manager, tesla_fsd_scene_fsd_running);
+            consumed = true;
+            break;
+        case MainMenuSettings:
+            scene_manager_next_scene(app->scene_manager, tesla_fsd_scene_settings);
             consumed = true;
             break;
         case MainMenuAbout: