+++ scenes_config/app_scene_config.h
@@ -1,4 +1,5 @@
 ADD_SCENE(tesla_fsd, main_menu, main_menu)
+ADD_SCENE(tesla_fsd, settings, settings)
 ADD_SCENE(tesla_fsd, hw_detect, hw_detect)
 ADD_SCENE(tesla_fsd, hw_select, hw_select)
 ADD_SCENE(tesla_fsd, fsd_running, fsd_running)